Fixes due to renaming .cXX to .estXX.

+++ code/demo/romabundledemo.m
@@ -86,7 +86,7 @@ plotopstats(result{1},E{1},COP);
 
 for i=1:length(E)
     fig=tagfigure(sprintf('network%d',i));
-    fprintf('Displaying bundle iteration playback for method %s in figure %d.\n',E{i}.damping.name,fig);
+    fprintf('Displaying bundle iteration playback for method %s in figure %d.\n',E{i}.damping.name,double(fig));
     plotnetwork(result{i},E{i},'trans','up','align',1,'title',...
                 ['Damping: ',dampings{i},'. Iteration %d of %d'], ...
                 'axes',fig,'pause','on');
